review AI-gen code (wip)
