+++ src/modules/admin/components/ceMatchRules/AffectedUnitGroupsTable.tsx
@@ -1,10 +1,9 @@
-import { Box, Typography } from '@mui/material';
+import { Box } from '@mui/material';
 import pluralize from 'pluralize';
 import { useMemo } from 'react';
 
 import GenericTable from '@/components/elements/table/GenericTable';
 import { ColumnDef } from '@/components/elements/table/types';
-import { WarningSection } from '@/modules/errors/components/WarningAlert';
 
 // Expected shape of the items returned from the server in `errors[x].data.affectedUnitGroups`.
 export interface AffectedUnitGroup {
@@ -55,17 +54,11 @@ const AffectedUnitGroupsTable = ({
 
   return (
     <Box sx={{ mt: 2 }}>
-      <WarningSection
-        header={<Typography fontWeight={600}>Affected unit groups</Typography>}
-      >
-        <Box sx={{ mx: -2, mb: -2 }}>
-          <GenericTable
-            rows={unitGroups}
-            columns={columns}
-            rowSx={() => ({ td: { py: 1.5 } })}
-          />
-        </Box>
-      </WarningSection>
+      <GenericTable
+        rows={unitGroups}
+        columns={columns}
+        rowSx={() => ({ td: { py: 1 } })}
+      />
     </Box>
   );
 };
